+++ src/resources/views/partials/main/exception_line.blade.php
@@ -1,0 +1,4 @@
+<tr>
+    <td><code>{!! $exception->value !!}</code></td>
+    <td>{!! $exception->description !!}</td>
+</tr>